updated readme
Only includes base goals and definitely is a starting point, but this
time I clearly know where we're headed and what we're trying to do.
When I'm less tired, I can go through old documents and conversations
with George and others to add more meat to the specs and goals.